% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{auth_token}
\alias{auth_token}
\title{Return the currently configured API authentication token}
\usage{
auth_token()
}
\value{
API token
}
\description{
Used by internal functions to obtain auth token required
to make API requests. Will fail with an user error to
educate users on how to configure their token.
}
