% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_summary_response.R
\name{check_dataset_summary_response}
\alias{check_dataset_summary_response}
\title{Validate \code{get_dataset} response object.}
\usage{
check_dataset_summary_response(object)
}
\arguments{
\item{object}{Object of type \code{\link{dataset_summary_response}}.}
}
\value{
TRUE/FALSE based on validity of object.
}
\description{
Validate \code{get_dataset} response object.
}
