% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_summary_response.R
\name{check_file_summary_response}
\alias{check_file_summary_response}
\title{Validate \code{get_dataset} response files.}
\usage{
check_file_summary_response(object)
}
\arguments{
\item{object}{Object of type \code{\link{file_summary_response}}}
}
\value{
TRUE/FALSE depending on validity of object.
}
\description{
Validate \code{get_dataset} response files.
}
