% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight_summary_response.R
\name{check_insight_summary_response}
\alias{check_insight_summary_response}
\title{Validate \code{insight_summary_response} object.}
\usage{
check_insight_summary_response(object)
}
\arguments{
\item{object}{Object of type \code{\link{insight_summary_response}}.}
}
\description{
Validate \code{insight_summary_response} object.
}
