% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_summary_response.R
\name{check_project_summary_response}
\alias{check_project_summary_response}
\title{Validate \code{get_project} response object.}
\usage{
check_project_summary_response(object)
}
\arguments{
\item{object}{Object of type \code{\link{project_summary_response}}.}
}
\description{
Validate \code{get_project} response object.
}
