% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_info_response.R
\name{check_user_info_response}
\alias{check_user_info_response}
\title{Validate \code{get_user} response object, returning the object if valid, and stopping with an error message if invalid.}
\usage{
check_user_info_response(object)
}
\arguments{
\item{object}{Object of type \code{\link{user_info_response}}.}
}
\value{
the object
}
\description{
Validate \code{get_user} response object, returning the object if valid, and stopping with an error message if invalid.
}
