% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{configure}
\alias{configure}
\title{Library configuration tool.}
\usage{
configure(auth_token = NULL)
}
\arguments{
\item{auth_token}{data.world's API authentication token.}
}
\description{
Configuration is not persistent and must be performed for
every new R session.
}
\section{DO NOT SHARE YOUR AUTHENTICATION TOKEN}{


For your security, do not include your API authentication token in code that
is intended to be shared with others.

Call this function via console, always when possible.

If you must call it in code do not include the actual API token.
Instead, pass the token via a variable in .Renviron, and do not share
your .Renviron file. For example:

\code{
  dwapi::configure(auth_token = Sys.getenv("DW_AUTH_TOKEN"))
}
}

\examples{
dwapi::configure(auth_token = "YOUR_API_TOKEN_HERE")
}
