% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset_response}
\alias{create_dataset_response}
\title{De-serialize a structured list into create_dataset_reponse object.}
\usage{
create_dataset_response(structure)
}
\arguments{
\item{structure}{httr response object.}
}
\value{
Object of type \code{\link{create_dataset_response}}.
}
\description{
De-serialize a structured list into create_dataset_reponse object.
}
