% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_insight.R
\name{create_insight}
\alias{create_insight}
\title{Create an insight.}
\usage{
create_insight(project_owner, project_id, create_insight_req)
}
\arguments{
\item{project_owner}{ID of project owner}

\item{project_id}{ID of project to which insight is to be added}

\item{create_insight_req}{Request object of
type \code{\link{insight_create_request}}.}
}
\value{
Object of type \code{\link{create_insight_response}}.
}
\description{
Create an insight.
}
\examples{
\dontrun{
  dwapi::create_insight(
    project_owner = 'user',
    project_id = 'project',
    insight_create_request(title='My Insight', image_url='https://...'))
}
}
