% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_insight.R
\name{create_insight_response}
\alias{create_insight_response}
\title{De-serialize a structured list into create_insight_response object.}
\usage{
create_insight_response(structure)
}
\arguments{
\item{structure}{httr response object.}
}
\value{
Object of type \code{\link{create_insight_response}}.
}
\description{
De-serialize a structured list into create_insight_response object.
}
\keyword{internal}
