% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_summary_response.R
\name{dataset_summary_response}
\alias{dataset_summary_response}
\title{Deserialize \code{get_dataset} response object.}
\usage{
dataset_summary_response(structure)
}
\arguments{
\item{structure}{httr response object.}
}
\value{
Object of type \code{\link{dataset_summary_response}}.
}
\description{
Deserialize \code{get_dataset} response object.
}
\seealso{
\code{\link{get_dataset}}
}
