% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_dataset.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Delete a dataset}
\usage{
delete_dataset(dataset_owner, dataset_id)
}
\arguments{
\item{dataset_owner}{ID of dataset owner}

\item{dataset_id}{ID of dataset to delete}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Delete a dataset
}
\examples{
\dontrun{
  dwapi::delete_dataset(
    dataset_owner = 'user',
    dataset_id = 'project')
}
}
