% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_file.R
\name{delete_file}
\alias{delete_file}
\title{Delete a single file from a dataset.}
\usage{
delete_file(owner_id, dataset_id, file_name)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}

\item{file_name}{File name, including the file extension.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Delete a single file from a dataset.
}
\examples{
\dontrun{
  dwapi::delete_file('user', 'dataset',
    file_name = 'file.csv')
}
}
