% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_file.R
\name{delete_files}
\alias{delete_files}
\title{Delete one or more files from a dataset.}
\usage{
delete_files(owner_id, dataset_id, file_names)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}

\item{file_names}{List of file names, including file extensions.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Delete one or more files from a dataset.
}
\examples{
\dontrun{
  dwapi::delete_files('user', 'dataset',
    file_names = list('file1.csv', 'file2.csv'))
}
}
