% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_insight.R
\name{delete_insight}
\alias{delete_insight}
\title{Delete an insight.}
\usage{
delete_insight(project_owner, project_id, id)
}
\arguments{
\item{project_owner}{ID of project owner}

\item{project_id}{ID of project of which insight is a component}

\item{id}{ID of insight to be replaced}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Delete an insight.
}
\examples{
\dontrun{
  dwapi::delete_insight(
    project_owner = 'user',
    project_id = 'project',
    id = 'insight')
}
}
