% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_project.R
\name{delete_project}
\alias{delete_project}
\title{Delete a project.}
\usage{
delete_project(project_owner, project_id)
}
\arguments{
\item{project_owner}{ID of project owner}

\item{project_id}{ID of project to delete}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Delete a project.
}
\examples{
\dontrun{
  dwapi::delete_project(
    project_owner = 'user',
    project_id = 'project')
}
}
