% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datapackage.R
\name{download_datapackage}
\alias{download_datapackage}
\title{Download data package.}
\usage{
download_datapackage(owner_id, dataset_id, output_path = NULL)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}

\item{output_path}{Directory path, where data package will be saved.}
}
\description{
A data package is a portable dataset file format.
Learn more about the Data Package specification at
\url{http://frictionlessdata.io/data-packages/}.
}
\examples{
\dontrun{
  dwapi::download_datapackage(
    "user", "dataset",
    output_path = tempdir())
}
}
