% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download a dataset as a zip archive and store at the specified path.}
\usage{
download_dataset(dataset, output_path)
}
\arguments{
\item{dataset}{Dataset URL or path.}

\item{output_path}{Local file path where dataset zip file will be saved.}
}
\value{
Server response message.
}
\description{
Download a dataset as a zip archive and store at the specified path.
}
\examples{
\dontrun{
  dwapi::download_dataset(dataset = 'user/dataset',
   output_path = tempfile(fileext = 'zip'))
}
}
