% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download file from dataset onto the local file system.}
\usage{
download_file(owner_id, dataset_id, file_name, output)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}

\item{file_name}{File name, including file extension.}

\item{output}{Local file path, where dataset file will be saved.}
}
\value{
Server response message.
}
\description{
Download file from dataset onto the local file system.
}
\examples{
\dontrun{
  dwapi::download_file('user', 'dataset',
   file_name = 'file.csv', output = tempfile(fileext = 'csv'))
}
}
