% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_table_as_data_frame.R
\name{download_table_as_data_frame}
\alias{download_table_as_data_frame}
\title{Download a dataset table onto a data frame.}
\usage{
download_table_as_data_frame(dataset, table_name)
}
\arguments{
\item{dataset}{Dataset URL or path.}

\item{table_name}{Table name.}
}
\value{
Data frame with data from table.
}
\description{
Download a dataset table onto a data frame.
}
\examples{
\dontrun{
  table_df <- dwapi::download_table_as_data_frame("user/dataset", "table")
}
}
\seealso{
\code{\link{list_tables}}
}
