% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_message.R
\name{error_message}
\alias{error_message}
\title{Deserialize error response object.}
\usage{
error_message(structure)
}
\arguments{
\item{structure}{httr response object}
}
\value{
Object of type \code{\link{error_message}}
}
\description{
Deserialize error response object.
}
