% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_project_key}
\alias{extract_project_key}
\title{Extract the project key from URL or as provided}
\usage{
extract_project_key(tentative_key)
}
\arguments{
\item{tentative_key}{key or URL}
}
\value{
project key extracted form URL or as provided
}
\description{
Extract the project key from URL or as provided
}
