% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_batch_update_request.R
\name{file_batch_update_request}
\alias{file_batch_update_request}
\title{Create request object for adding or updating dataset files.}
\usage{
file_batch_update_request(file_sources = list())
}
\arguments{
\item{file_sources}{(optional) List of \code{\link{file_create_or_update_request}} objects.}
}
\value{
Object of type \code{file_batch_update_request}.
}
\description{
Create request object for adding or updating dataset files.
}
\examples{
request <- dwapi::file_batch_update_request()
request <- dwapi::add_file(
  request = request, name = 'file.csv', url = 'http://data.world/file.csv')
}
\seealso{
\code{\link{add_files_by_source}}, \code{\link{add_file}}
}
