% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_source_create_request.R
\name{file_source_create_request}
\alias{file_source_create_request}
\title{Create object for adding dataset file sources.}
\usage{
file_source_create_request(url)
}
\arguments{
\item{url}{Source URL of file.}
}
\value{
Object of type \code{file_source_create_request}.
}
\description{
Create object for adding dataset file sources.
}
