% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_summary_response.R
\name{file_summary_response}
\alias{file_summary_response}
\title{Deserialize \code{get_dataset} response files.}
\usage{
file_summary_response(structure)
}
\arguments{
\item{structure}{httr response object.}
}
\value{
Object of type \code{file_summary_response}
}
\description{
Deserialize \code{get_dataset} response files.
}
