% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets_err}
\alias{get_datasets_err}
\title{Stop execution with a message when the response contains an error condition}
\usage{
get_datasets_err(response)
}
\arguments{
\item{response}{the response}
}
\description{
Stop execution with a message when the response contains an error condition
}
\keyword{internal}
