% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_library.R
\name{get_datasets_user_contributing}
\alias{get_datasets_user_contributing}
\title{Search for datasets contributed-to by the currently authenticated user.}
\usage{
get_datasets_user_contributing(limit = NULL, next_page_token = NULL)
}
\arguments{
\item{limit}{Maximum number of items to return}

\item{next_page_token}{Unique token used to retrieve next page}
}
\value{
a named list with at most two elements.  It will always
contain a list, named \code{records},
of objects of type \code{\link{dataset_summary_response}}. If the call
to \code{get_datasets_user_contributing()} was made with a non-null
\code{limit} parameter, and if further pages remain for retrieval, then
the list will also contain a single-element character vector,
named \code{next_page_token},
with the token to use in a subsequent call to get the next page.
}
\description{
Search for datasets contributed-to by the currently authenticated user.
}
\examples{
\dontrun{
  datasets_contributing <- dwapi::get_datasets_user_contributing()
}
}
