% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_insight.R
\name{get_insight}
\alias{get_insight}
\title{Retrieve an insight.}
\usage{
get_insight(project_owner, project_id, id)
}
\arguments{
\item{project_owner}{username of the owner of the project.}

\item{project_id}{identifier of the project.}

\item{id}{identifier of the insight}
}
\value{
an object of type \code{insight_summary_response}.
}
\description{
Retrieve an insight.
}
\examples{
\dontrun{
  dwapi::get_insights(project_owner = "user", project_id = "project_id")
}
}
