% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight_summary_response.R
\name{get_insight_optional_fields}
\alias{get_insight_optional_fields}
\title{Get the insight optional fields}
\usage{
get_insight_optional_fields()
}
\description{
Get the insight optional fields
}
\keyword{internal}
