% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight_summary_response.R
\name{get_insight_required_fields}
\alias{get_insight_required_fields}
\title{Get the insight required fields}
\usage{
get_insight_required_fields()
}
\description{
Get the insight required fields
}
\keyword{internal}
