% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{get_project}
\alias{get_project}
\title{Retrieve a project.}
\usage{
get_project(project_owner, project_id)
}
\arguments{
\item{project_owner}{username of the owner of the project.}

\item{project_id}{identifier of the project.}
}
\value{
an object of type \code{project_summary_response}.
}
\description{
Retrieve a project.
}
\examples{
\dontrun{
  dwapi::get_project(project_owner = "user", project_id = "project_id")
}
}
