% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_summary_response.R
\name{get_project_required_fields}
\alias{get_project_required_fields}
\title{Get the project required fields}
\usage{
get_project_required_fields()
}
\description{
Get the project required fields
}
\keyword{internal}
