% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_library.R
\name{get_projects_user_liked}
\alias{get_projects_user_liked}
\title{Search for projects liked by the currently authenticated user.}
\usage{
get_projects_user_liked(limit = NULL, next_page_token = NULL)
}
\arguments{
\item{limit}{Maximum number of items to return}

\item{next_page_token}{Unique token used to retrieve next page}
}
\value{
a named list with at most two elements.  It will always
contain a list, named \code{records},
of objects of type \code{\link{project_summary_response}}. If
the call to \code{get_projects_user_liked()} was made with a non-null
\code{limit} parameter, and if further pages remain for retrieval,
then
the list will also contain a single-element character vector,
named \code{next_page_token},
with the token to use in a subsequent call to get the next page.
}
\description{
Search for projects liked by the currently authenticated user.
}
\examples{
\dontrun{
  projects_liked <- dwapi::get_projects_user_liked()
}
}
