% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_schema.R
\name{get_table_schema}
\alias{get_table_schema}
\title{Retrieve schema information for a dataset table.}
\usage{
get_table_schema(owner_id, dataset_id, table_name)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}

\item{table_name}{Table name.}
}
\value{
Object of type \code{\link{table_schema_response}}.
}
\description{
Retrieve schema information for a dataset table.
}
\examples{
\dontrun{
  table_schema <- dwapi::get_table_schema("user", "dataset", "table")
}
}
\seealso{
\code{\link{list_tables}}
}
