% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user}
\alias{get_user}
\title{Retrieve user information for the currently authenticated user.}
\usage{
get_user()
}
\value{
Object of type \code{\link{user_info_response}}.
}
\description{
Retrieve user information for the currently authenticated user.
}
\examples{
\dontrun{
  my_dw_user <- dwapi::get_user()
}
}
