% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_user_assets}
\alias{get_user_assets}
\title{Search for assets owned by, liked by, or contributed-to by the currently authenticated user.}
\usage{
get_user_assets(type = c("datasets", "projects"), role = c("own", "liked",
  "contributing"), limit = NULL, nextPageToken = NULL, sort = NULL)
}
\arguments{
\item{type}{the type of asset (one of: datasets, projects)}

\item{role}{the user's role with respect to the asset (one of: own, liked, contributing)}
}
\value{
a named list with at most two elements.  It will always contain a list, named \code{records},
of objects of a list structure type appropriate for the kind of asset queried. If the call to \code{get_datasets()} was made with a non-null
\code{limit} parameter, and if further pages remain for retrieval, then
the list will also contain a single-element character vector, named \code{nextPageToken},
with the token to use in a subsequent call to get the next page.
}
\description{
Search for assets owned by, liked by, or contributed-to by the currently authenticated user.
}
\keyword{internal}
