% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_library.R
\name{handle_library_err}
\alias{handle_library_err}
\title{Stop execution with a message when the response contains an error condition}
\usage{
handle_library_err(response)
}
\arguments{
\item{response}{the response}
}
\description{
Stop execution with a message when the response contains an error condition
}
\keyword{internal}
