% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight_summary_response.R
\name{insight_summary_response}
\alias{insight_summary_response}
\title{Deserialize \code{get_insight} response object.}
\usage{
insight_summary_response(structure)
}
\arguments{
\item{structure}{httr response object.}
}
\value{
Object of type \code{\link{insight_summary_response}}.
}
\description{
Deserialize \code{get_insight} response object.
}
\seealso{
\code{\link{get_insights}} \code{\link{get_insight}}
}
