% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight_request.R
\name{insight_update_request}
\alias{insight_update_request}
\title{Create insight object for updating existing insights}
\usage{
insight_update_request(title = NULL, description = NULL, image_url = NULL,
  embed_url = NULL, markdown_body = NULL, source_link = NULL,
  data_source_links = NULL)
}
\arguments{
\item{title}{Insight title}

\item{description}{Optional insight description}

\item{image_url}{URL of image representing the insight}

\item{embed_url}{URL of content to embed}

\item{markdown_body}{Markdown text containing the insight}

\item{source_link}{Permalink to source code or platform this
insight was generated with}

\item{data_source_links}{List containing one or more permalinks
to the data sources used to generate this insight}
}
\value{
Request object of type \code{insight_replace_request}.
}
\description{
Create insight object for updating existing insights
}
\examples{
request <- dwapi::insight_update_request(title='New title')
request <- dwapi::insight_update_request(title='New title',
    description = 'New description')
}
\seealso{
\code{\link{create_insight}}
}
