% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.blank}
\alias{is.blank}
\title{Determine if a value has NA or zero-length character elements, or is NULL}
\usage{
is.blank(value)
}
\arguments{
\item{value}{the value to test}
}
\value{
a boolean vector of the same length as \code{value}, whose elements
are TRUE if the corresponding element in \code{value} is NA or zero-length
character, and FALSE otherwise. If \code{value} is NULL, then a singleton
TRUE vector is returned.
}
\description{
Determine if a value has NA or zero-length character elements, or is NULL
}
\keyword{internal}
