% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_dataset.R
\name{link_dataset}
\alias{link_dataset}
\title{Link a dataset to a project.}
\usage{
link_dataset(project_owner, project_id, dataset_owner, dataset_id)
}
\arguments{
\item{project_owner}{ID of project owner.}

\item{project_id}{ID of project to which the dataset is to be linked.}

\item{dataset_owner}{ID of dataset owner.}

\item{dataset_id}{ID of dataset to be linked.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Link a dataset to a project.
}
\examples{
\dontrun{
  dwapi::link_dataset(
    project_owner = 'user1',
    project_id = 'project',
    dataset_owner = 'user2',
    dataset_id = 'dataset')
}
}
