% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked_dataset_create_or_update_request.R
\name{linked_dataset_create_or_update_request}
\alias{linked_dataset_create_or_update_request}
\title{Create request object for new linked datasets, or updates to existing
datasets.}
\usage{
linked_dataset_create_or_update_request(owner, id)
}
\arguments{
\item{owner}{User name and unique identifier of the creator of the dataset.}

\item{id}{Unique identifier of dataset.}
}
\value{
Request object of type \code{linked_dataset_create_or_update_request}.
}
\description{
Create request object for new linked datasets, or updates to existing
datasets.
}
\examples{
request <- dwapi::linked_dataset_create_or_update_request(
  owner = 'ownerId', id = 'datasetId')
}
\seealso{
\code{\link{create_project}}
}
