% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tables.R
\name{list_tables}
\alias{list_tables}
\title{List tables in a dataset.}
\usage{
list_tables(owner_id, dataset_id)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}
}
\value{
list of table names.
}
\description{
Tables are a logical concept representing normalized data extracted
from tabular files uploaded to datasets on data.world.
}
\details{
\strong{EXPERIMENTAL}: This is an experimental feature and
backwards-compability is not guaranteed in future releases.
}
\examples{
\dontrun{
  tables <- dwapi::list_tables("user", "dataset")
}
}
