% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_success_or_error}
\alias{parse_success_or_error}
\title{Parse simple responses (success or error).}
\usage{
parse_success_or_error(response)
}
\arguments{
\item{response}{httr response.}
}
\value{
Deserialized success or error.
}
\description{
Parse simple responses (success or error).
}
