% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_summary_response.R
\name{project_summary_response}
\alias{project_summary_response}
\title{Deserialize \code{get_project} response object.}
\usage{
project_summary_response(structure)
}
\arguments{
\item{structure}{httr response object.}
}
\value{
Object of type \code{\link{project_summary_response}}.
}
\description{
Deserialize \code{get_project} response object.
}
\seealso{
\code{\link{get_projects_user_own}}
}
