% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_insight.R
\name{replace_insight}
\alias{replace_insight}
\title{Replace an insight.}
\usage{
replace_insight(project_owner, project_id, id, replace_insight_req)
}
\arguments{
\item{project_owner}{ID of project owner}

\item{project_id}{ID of project to which insight is to be added}

\item{id}{ID of insight to be replaced}

\item{replace_insight_req}{Request object of
type \code{\link{insight_replace_request}}.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Replace an insight.
}
\examples{
\dontrun{
  dwapi::replace_insight(
    project_owner = 'user',
    project_id = 'project',
    id = 'insight',
    insight_replace_request(title='My Insight', image_url='https://...'))
}
}
