% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_project.R
\name{replace_project_response}
\alias{replace_project_response}
\title{De-serialize a structured list into create_project_reponse object.}
\usage{
replace_project_response(structure)
}
\arguments{
\item{structure}{httr response object.}
}
\value{
Object of type \code{\link{create_project_response}}.
}
\description{
De-serialize a structured list into create_project_reponse object.
}
