% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql}
\alias{sql}
\title{Execute SQL query against a dataset.}
\usage{
sql(owner_id, dataset_id, query, query_params = list())
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}

\item{query}{SQL query.}

\item{query_params}{List of positional query parameters.}
}
\value{
Data frame with data from query results.
}
\description{
Execute SQL query against a dataset.
}
\examples{
\dontrun{
  dwapi::sql("user", "dataset",
    query="SELECT *
           FROM TableName
           LIMIT 10")

  dwapi::sql(dataset="user/dataset",
    query="SELECT *
           FROM TableName where `field1` = ? AND `field2` > ?
           LIMIT 10",
    queryParameters = list("value", 5.0))
}
}
