% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_insights.R
\name{stop_on_insights_error}
\alias{stop_on_insights_error}
\title{Handle a response containing an error by stopping execution with the
embedded message}
\usage{
stop_on_insights_error(response)
}
\description{
Handle a response containing an error by stopping execution with the
embedded message
}
\keyword{internal}
