% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_message.R
\name{success_message}
\alias{success_message}
\title{Deserialize success response object.}
\usage{
success_message(structure)
}
\arguments{
\item{structure}{httr response object}
}
\value{
Object of type \code{\link{success_message}}
}
\description{
Deserialize success response object.
}
