% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_dataset.R
\name{sync}
\alias{sync}
\title{Fetch latest files from source and update dataset.}
\usage{
sync(owner_id, dataset_id)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Fetch latest files from source and update dataset.
}
\examples{
\dontrun{
  dwapi::sync('user', 'dataset')
}
}
