% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_schema_response.R
\name{table_schema_field_response}
\alias{table_schema_field_response}
\title{Deserialize \code{get_table_schema} fields.}
\usage{
table_schema_field_response(structure)
}
\arguments{
\item{structure}{httr response object}
}
\value{
Object of type \code{table_schema_field_response}
}
\description{
Deserialize \code{get_table_schema} fields.
}
