% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_schema_update_request.R
\name{table_schema_field_update_request}
\alias{table_schema_field_update_request}
\title{Create request object for updating table schema fields.}
\usage{
table_schema_field_update_request(name, description)
}
\arguments{
\item{name}{Table field name.}

\item{description}{Table field description.}
}
\value{
Object of type \code{table_schema_field_update_request}
}
\description{
Create request object for updating table schema fields.
}
\examples{
field_update_req <- dwapi::table_schema_field_update_request("field", "new desc")
}
