% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_schema_update_request.R
\name{table_schema_update_request}
\alias{table_schema_update_request}
\title{Create request object for updating table schema.}
\usage{
table_schema_update_request(fields)
}
\arguments{
\item{fields}{List of \code{\link{table_schema_field_update_request}} objects.}
}
\value{
Object of type \code{table_schema_update_request}.
}
\description{
Create request object for updating table schema.
}
\examples{
field_update_req <- dwapi::table_schema_field_update_request("field", "new desc")
schema_update_req <- dwapi::table_schema_update_request(c(field_update_req))
}
\seealso{
\code{\link{update_table_schema}}
}
