% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlink_dataset.R
\name{unlink_dataset}
\alias{unlink_dataset}
\title{Unlink a dataset from a project.}
\usage{
unlink_dataset(project_owner, project_id, dataset_owner, dataset_id)
}
\arguments{
\item{project_owner}{ID of project owner.}

\item{project_id}{ID of project from which the dataset is to be unlinked.}

\item{dataset_owner}{ID of dataset owner.}

\item{dataset_id}{ID of dataset to be unlinked.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Unlink a dataset from a project.
}
\examples{
\dontrun{
  dwapi::unlink_dataset(
    project_owner = 'user1',
    project_id = 'project',
    dataset_owner = 'user2',
    dataset_id = 'dataset')
}
}
