% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dataset.R
\name{update_dataset}
\alias{update_dataset}
\title{Update an existing dataset.}
\usage{
update_dataset(owner_id, dataset_id, dataset_update_req)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}

\item{dataset_update_req}{Request object of type \code{\link{dataset_update_request}}.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Update an existing dataset.
}
\examples{
request <- dwapi::dataset_update_request(visibility = 'OPEN',
  description = 'UPDATED DESCRIPTION !')

\dontrun{
  dwapi::update_dataset('user', 'dataset', dataset_update_req = request)
}
}
