% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_insight.R
\name{update_insight}
\alias{update_insight}
\title{Update an insight.}
\usage{
update_insight(project_owner, project_id, id, update_insight_req)
}
\arguments{
\item{project_owner}{ID of project owner}

\item{project_id}{ID of project of which insight is a component}

\item{id}{ID of insight to be replaced}

\item{update_insight_req}{Request object of
type \code{\link{insight_update_request}}.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Update an insight.
}
\examples{
\dontrun{
  dwapi::update_insight(
    project_owner = 'user',
    project_id = 'project',
    id = 'insight',
    insight_update_request(title='My Insight', image_url='https://...'))
}
}
