% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_project.R
\name{update_project}
\alias{update_project}
\title{Update an existing project.}
\usage{
update_project(owner_id, project_id, project_update_req)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{project_id}{Project unique identifier}

\item{project_update_req}{Request object of type \code{\link{project_update_request}}.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Update an existing project.
}
\examples{
request <- dwapi::project_update_request(
  objective = 'UPDATED OBJECTIVE !')

\dontrun{
  dwapi::update_project('user', 'project', project_update_req = request)
}
}
