% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_table_schema.R
\name{update_table_schema}
\alias{update_table_schema}
\title{Update table schema.}
\usage{
update_table_schema(owner_id, dataset_id, table_name, table_schema_update_req)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}

\item{table_name}{Table name.}

\item{table_schema_update_req}{Request object of
type \code{\link{table_schema_update_request}}}
}
\value{
Object of type \code{\link{success_message}}
}
\description{
Update table schema.
}
\examples{
field_update_req <- dwapi::table_schema_field_update_request(
"field", "new desc")
schema_update_req <- dwapi::table_schema_update_request(c(field_update_req))
\dontrun{
  dwapi::update_table_schema("user", "dataset", "table", schema_update_req)
}
}
