% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_file}
\alias{upload_file}
\title{Upload a single file to a dataset or project.}
\usage{
upload_file(owner_id, dataset_id, path, file_name)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Unique identifier of dataset or project}

\item{path}{File path on local file system.}

\item{file_name}{File name, including file extension.}
}
\value{
Server response message.
}
\description{
Upload a single file to a dataset or project.
}
\examples{
\dontrun{
  dwapi::upload_file('user', 'dataset', file_name = 'file.csv',
    path = 'file.csv')
}
}
