% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_info_response.R
\name{user_info_response}
\alias{user_info_response}
\title{Deserialize \code{get_user} response object.}
\usage{
user_info_response(structure)
}
\arguments{
\item{structure}{httr response object.}
}
\value{
Object of type \code{\link{user_info_response}}.
}
\description{
Deserialize \code{get_user} response object.
}
\seealso{
\code{\link{get_dataset}}
}
