% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_dataset.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Delete a dataset}
\usage{
delete_dataset(owner_id, dataset_id)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Delete a dataset
}
\examples{
\dontrun{
  dwapi::delete_dataset('user', 'dataset')
}
}
