% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_dataset.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Delete a dataset}
\usage{
delete_dataset(dataset)
}
\arguments{
\item{dataset}{Dataset URL or path.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Delete a dataset
}
\examples{
\dontrun{
  dwapi::delete_dataset(
    dataset = 'user/dataset')
}
}
