% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_file.R
\name{delete_files}
\alias{delete_files}
\title{Delete one or more files from a dataset.}
\usage{
delete_files(dataset, file_names)
}
\arguments{
\item{dataset}{Dataset URL or path.}

\item{file_names}{List of file names, including file extensions.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Delete one or more files from a dataset.
}
\examples{
\dontrun{
  dwapi::delete_files(dataset = 'user/dataset',
    file_names = list('file1.csv', 'file2.csv'))
}
}
