% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datapackage.R
\name{download_datapackage}
\alias{download_datapackage}
\title{Download data package.}
\usage{
download_datapackage(dataset, output_path = NULL)
}
\arguments{
\item{dataset}{Dataset URL or path.}

\item{output_path}{Directory path, where data package will be saved.}
}
\description{
A data package is a portable dataset file format.
Learn more about the Data Package specification at
\url{http://frictionlessdata.io/data-packages/}.
}
\examples{
\dontrun{
  dwapi::download_datapackage(
    dataset = "user/dataset",
    output_path = "tmp/user/dataset")
}
}
