% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download a dataset as a zip archive and store at the specified path.}
\usage{
download_dataset(owner_id, dataset_id, output_path)
}
\arguments{
\item{owner_id}{User name and unique identifier of the creator of a
dataset or project}

\item{dataset_id}{Dataset unique identifier}

\item{output_path}{Local file path where dataset zip file will be saved.}
}
\value{
Server response message.
}
\description{
Download a dataset as a zip archive and store at the specified path.
}
\examples{
\dontrun{
  dwapi::download_dataset('user', 'dataset',
   output_path = tempfile(fileext = 'zip'))
}
}
