% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download file from dataset onto the local file system.}
\usage{
download_file(dataset, file_name, output)
}
\arguments{
\item{dataset}{Dataset URL or path.}

\item{file_name}{File name, including file extension.}

\item{output}{Local file path, where dataset file will be saved.}
}
\value{
Server response message.
}
\description{
Download file from dataset onto the local file system.
}
\examples{
\dontrun{
  download_file(dataset = 'user/dataset', file_name = 'file.csv',
    output = 'tmp/file.csv')
}
}
