% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Retrieve dataset information.}
\usage{
get_dataset(dataset)
}
\arguments{
\item{dataset}{Dataset URL or path.}
}
\value{
Object of type \code{\link{dataset_summary_response}}.
}
\description{
Retrieve dataset information.
}
\examples{
\dontrun{
  get_dataset(dataset = "user/dataset")
}
}
