% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_dataset.R
\name{sync}
\alias{sync}
\title{Fetch latest files from source and update dataset.}
\usage{
sync(dataset)
}
\arguments{
\item{dataset}{Dataset URL or path.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Fetch latest files from source and update dataset.
}
\examples{
\dontrun{
  dwapi::sync(dataset='user/dataset')
}
}
