% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_project.R
\name{update_project}
\alias{update_project}
\title{Update an existing project.}
\usage{
update_project(project, project_update_req)
}
\arguments{
\item{project}{Project URL or path.}

\item{project_update_req}{Request object of type \code{\link{project_update_request}}.}
}
\value{
Object of type \code{\link{success_message}}.
}
\description{
Update an existing project.
}
\examples{
request <- dwapi::project_update_request(
  objective = 'UPDATED OBJECTIVE !')

\dontrun{
  dwapi::update_project(project_update_req = request,
    dataset = 'user/project')
}
}
