% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data_frame.R
\name{upload_data_frame}
\alias{upload_data_frame}
\title{Upload a data frame as a file to a dataset.}
\usage{
upload_data_frame(dataset, data_frame, file_name)
}
\arguments{
\item{dataset}{Dataset URL or path.}

\item{data_frame}{Data frame object.}

\item{file_name}{File name, including file extension.}
}
\value{
Server response message.
}
\description{
Upload a data frame as a file to a dataset.
}
\examples{
df = data.frame(a = c(1,2,3),b = c(4,5,6))
\dontrun{
  upload_data_frame(file_name = 'sample.csv',
    data_frame = df, dataset = 'user/dataset')
}
}
