% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_files}
\alias{upload_files}
\title{Upload one or more files to a dataset.}
\usage{
upload_files(dataset, paths)
}
\arguments{
\item{dataset}{Dataset URL or path.}

\item{paths}{List of file paths on local file system.}
}
\value{
Server response message.
}
\description{
Upload one or more files to a dataset.
}
\examples{
\dontrun{
  dwapi::upload_files(dataset = 'user/dataset',
    paths = c('file1.csv', 'file2.csv'))
}
}
